connection statistics bytes and packet counters changed from 32 to 64 bit long integers

+++ Lidgren.Network/NetConnectionStatistics.cs
@@ -40,18 +40,18 @@ namespace Lidgren.Network
 	{
 		private readonly NetConnection m_connection;
 
-		internal int m_sentPackets;
-		internal int m_receivedPackets;
+		internal long m_sentPackets;
+		internal long m_receivedPackets;
 
-		internal int m_sentMessages;
-		internal int m_receivedMessages;
-		internal int m_receivedFragments;
+		internal long m_sentMessages;
+		internal long m_receivedMessages;
+		internal long m_receivedFragments;
 
-		internal int m_sentBytes;
-		internal int m_receivedBytes;
+		internal long m_sentBytes;
+		internal long m_receivedBytes;
 
-		internal int m_resentMessagesDueToDelay;
-		internal int m_resentMessagesDueToHole;
+		internal long m_resentMessagesDueToDelay;
+		internal long m_resentMessagesDueToHole;
 
 		internal NetConnectionStatistics(NetConnection conn)
 		{
@@ -75,27 +75,27 @@ namespace Lidgren.Network
 		/// <summary>
 		/// Gets the number of sent packets for this connection
 		/// </summary>
-		public int SentPackets { get { return m_sentPackets; } }
+		public long SentPackets { get { return m_sentPackets; } }
 
 		/// <summary>
 		/// Gets the number of received packets for this connection
 		/// </summary>
-		public int ReceivedPackets { get { return m_receivedPackets; } }
+		public long ReceivedPackets { get { return m_receivedPackets; } }
 
 		/// <summary>
 		/// Gets the number of sent bytes for this connection
 		/// </summary>
-		public int SentBytes { get { return m_sentBytes; } }
+		public long SentBytes { get { return m_sentBytes; } }
 
 		/// <summary>
 		/// Gets the number of received bytes for this connection
 		/// </summary>
-		public int ReceivedBytes { get { return m_receivedBytes; } }
+		public long ReceivedBytes { get { return m_receivedBytes; } }
 
 		/// <summary>
 		/// Gets the number of resent reliable messages for this connection
 		/// </summary>
-		public int ResentMessages { get { return m_resentMessagesDueToHole + m_resentMessagesDueToDelay; } }
+		public long ResentMessages { get { return m_resentMessagesDueToHole + m_resentMessagesDueToDelay; } }
 
 		// public double LastSendRespondedTo { get { return m_connection.m_lastSendRespondedTo; } }
